Added express-validator package

+++ server/controllers/indexController.js
@@ -1,4 +1,5 @@
 const asyncHandler = require('express-async-handler');
+const { body, validationResult } = require('express-validator');
 const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
 const { DateTime } = require('luxon');
@@ -54,28 +55,44 @@ exports.end_game_get = asyncHandler(async (req, res, next) => {
 });
 
 // Handles creating new user if score is top 10
-exports.end_game_post = asyncHandler(async (req, res, next) => {
-  const newUser = await prisma.user.create({
-    data: {
-      username: req.body.username,
-      imageId: req.body.imageId,
-      minutes: req.body.minutes,
-      seconds: req.body.seconds,
-    },
-  });
+exports.end_game_post = [
+  body('username')
+    .trim()
+    .isAlphanumeric()
+    .withMessage('Letters and numbers only')
+    .escape(),
 
-  await prisma.data.delete({
-    where: {
-      user_id: req.session.user,
-    },
-  });
-  req.session.destroy();
+  asyncHandler(async (req, res, next) => {
+    const errors = validationResult(req);
+    const newUser = await prisma.user.create({
+      data: {
+        username: req.body.username,
+        imageId: req.body.imageId,
+        minutes: req.body.minutes,
+        seconds: req.body.seconds,
+      },
+    });
 
-  res.json({
-    message: 'Game ended',
-    newUser
-  });
-});
+    if (!errors.isEmpty()) {
+      res.json({
+        errors: errors.array(),
+      });
+      return;
+    }
+
+    await prisma.data.delete({
+      where: {
+        user_id: req.session.user,
+      },
+    });
+    req.session.destroy();
+
+    res.json({
+      message: 'Game ended',
+      newUser,
+    });
+  }),
+];
 
 // Handles ending the game when the user quits or exits
 exports.quit_game_delete = asyncHandler(async (req, res, next) => {
